Adds PGBOUNCER_CONNECTION_RETRY default no
stunnel previously was configured to always retry the connection to the
database server. When transient network errors interrupt the connection,
stunnel should exit so that Heroku's platform may replace the process in
a runtime with better network connectivity. This commit makes stunnel
retry configurable and defaults it to not retry.