Preserve connection options at end of string with dbname
Appending connction options to the end of DATABASE_URL is discouraged,
as it could just not be allowed by Heroku, if the config var is managed
by an addon, like Heroku Postgres. It is not good to rely on this
configuration method in other environments.

However, d591f7f uncovered a bug in the parser regex, where the db name
could land in $5 or $6 if there were connection options. As $5 was empty
and the remaining parsing used whitespace as separators, DB_NAME was
then passed on as the db name + connection options. As colons are now
used as separators, this bug was noticed, but existing behavior where
connection options are appended to the db name is preserved, as there
could be an app out there that could have depended on this bug.