Change default pool size to 2
I have a customer who was running 20 unicorn workers on a performance L dyno. They were using default settings and wondering why database time was so slow.

I'm proposing we set this to a slightly higher default of 2.

In the long term I propose we pull some internal data to better set this default value.

I.e. collect the maximum database connections used by all apps on the platform that are using this buildpack as well as the max number of dynos and whether or not they're using preboot.

Another question is: is there some metric we could provide that would help customers set this value? I.e. time spent waiting on a getting a pgbouncer connection?