[changelog skip] Fix Escaping in Changelog Script (#127)
* [changelog skip] Ensure PRs include a Changelog entry

The goal of this PR is to add a github action that checks for the presence of a changelog entry.

It is better to add entries as a PR is merged instead of having to remember what was merged and generate a changelog at release time.

By automating this check, it's one less thing the maintainer has to remember, and it's one less thing a change might be blocked on i.e. "Looks good, but please add a changelog entry".

Let me know if you have any questions and Happy Friday!

* [changelog skip] Fix Escaping in Changelog Script

The previous PR had a bug where the REGEX for grep was not properly escaped. This PR fixes that issue.

* Update check_changelog.yml