test: add Heroku stack testing
This commit adds a Dockerfile to build and run the buildpack in a
Heroku-like environment to test for compability, and actually execute
PgBouncer. It connects over a Unix socket to the management DB and
checks the currently running version, then exits.

Additional changes are made to `bin/compile` and `bin/start-pgbouncer`
to support this testing.

Fixes #165